# Improve performance by making use of C++11 movement.