--- a/kiwi/constraint.h
+++ b/kiwi/constraint.h
@@ -36,6 +36,8 @@ public:
 
     Constraint(const Constraint &other, double strength) : m_data(new ConstraintData(other, strength)) {}
 
+    Constraint(Constraint &&) = default;
+
     ~Constraint() {}
 
     const Expression &expression() const