--- a/kiwi/variable.h
+++ b/kiwi/variable.h
@@ -26,7 +26,7 @@ public:
 
     Variable(Context *context = 0) : m_data(new VariableData("", context)) {}
 
-    Variable(const std::string &name, Context *context = 0) : m_data(new VariableData(name, context)) {}
+    Variable(std::string name, Context *context = 0) : m_data(new VariableData(std::move(name), context)) {}
 
     Variable(const char *name, Context *context = 0) : m_data(new VariableData(name, context)) {}
 
@@ -86,8 +86,8 @@ private:
     {
 
     public:
-        VariableData(const std::string &name, Context *context) : SharedData(),
-                                                                  m_name(name),
+        VariableData(std::string name, Context *context) : SharedData(),
+                                                                  m_name(std::move(name)),
                                                                   m_context(context),
                                                                   m_value(0.0) {}