Hi!

**Edit:** after using google/benchmark on a dedicated Linux server, it seems this PR doesn't do anything relative to `suggestValue` usage and the values below are just some noise.
**However** this PR does help a bit with solver building performance (which is not part of the python benchmark) and it still relevant, see #91.

C++11 introduced movement semantic, which can be used to prevent copying and reallocating vectors and just pass them instead, this does help with performance a bit.

I have consistent benchmark time, with
```
kiwi.suggestValue: Mean +- std dev: 8.83 us +- 0.10 us
kiwi.suggestValue: Mean +- std dev: 8.88 us +- 0.10 us
kiwi.suggestValue: Mean +- std dev: 8.92 us +- 0.12 us
```

before this change and

```
kiwi.suggestValue: Mean +- std dev: 8.52 us +- 0.08 us
kiwi.suggestValue: Mean +- std dev: 8.60 us +- 0.15 us
kiwi.suggestValue: Mean +- std dev: 8.58 us +- 0.14 us
```

after.

It's not much but I think it's worth it :)