@Qix- 2020-09-05
Not sold on this. Compilers can do a good job at copy/move elision already. Explicit `std::move()` prevents a number of those optimizations, namely because `return std::move(x)` for example is not eligible for NRVO.

I'd _really_ like to see a comparison of compilers and perf benchmarks (perhaps checking Godbolt) before this is considered, as in practice a lot of `std:move()`'s are superfluous and can actually degrade performance.