@SirLynix 2020-09-05
The problem is, since copy constructor do have side-effects, the compiler just cannot optimize them as easily.

I'm aware of the problem with `return std::move(x);` which is why I never use it that way, but that's not what I did there, NRVO doesn't apply to parameters.

> defining a move constructor is OK, and in my experience is usually enough for the compiler to make the most appropriate choice (without the use of std::move()).
> ...
> as in practice a lot of std:move()'s are superfluous and can actually degrade performance

I clearly didn't have the same experience as you, as afaik the compiler just cannot choose, you're calling a constructor with side-effects and it has to call it most of the time.

Do you have a source for the fact that std::move calls are superfluous and actually degrade performance?

Here's some bench (+assembly) as requested: 

https://quick-bench.com/q/-RAcodG39scGYtmePVqga4WLBOw (Clang)
https://quick-bench.com/q/u2DMvbVRPx6n1qnkZ4qgwQ6vwUc (GCC)

As you can see, even in this simple example both compilers cannot optimize the vector copy (although it's a std type) if I don't use `std::move` and perform a copy.