@MatthieuDartiailh 2020-09-05
I must say I do not have a strong enough background in c++ to comment on the optimal solution here. However I would be curious to see what code you used to benchmark suggest value. I have a benchmark written in Python representative of a real world application of kiwi and if you can contribute a similar benchmark for the pure c++ version I would be interested. If there is a reasonable gain on this use case it would probably be worth merging the changes.

However another thing I need to think about is that it seems some of this changes change the API. Correct me if I am wrong but now when creating a constraint from an expression the original expression is destroyed, and similarly for the variable in a term. Similarly the signature of the Term constructor changed