@SirLynix 2020-09-05
I used your benchmark, the only difference is that i used pyperf instead of perf because I could'nt find a way to install it.

This does not create any API incompatibility, all "destroyed value" is only done in a movement context, you have to move your vector to give it to the constructor, if you don't it'll be copied, as before.