@SirLynix 2020-09-05
About the signature of the constructor, yes it did change but won't break any code as you don't have to change a thing on the caller side for this to work like before.