@SirLynix 2020-09-05
About making a C++ version of the benchmark, I'll look at that. Expect a third PR soon 😄