@MatthieuDartiailh 2020-09-09
@Qix- do you mind doing an extra review pass. To me it looks all good but you have obviously more C++ expertise than I do.