@SirLynix 2020-09-10
#91 shows that in fact this PR doesn't improve performance for `suggestValue`, it seems the consistent results I had were just some noise.

However, there's a real performance impact on solver building (which is logical as it's the place where constraint/others copying occurrs).