@MatthieuDartiailh 2020-09-11
Since all the C++ experts now seem to agree I will merge as soon as the conflicts have been handled.