@SirLynix 2020-09-11
Here you are, someone warned me I missed some noexcept qualifier and that adding it to default movement constructor/assignment operator would be better (as it would trigger a warning/error if it couldn't be satisfied). Having a noexcept constructor may be a requirement to allow movement in some rare cases (for example std::vector will prefer to copy its content when resizing rather than move it if the movement constructor is not noexcept).

See [std::vector::resize](https://en.cppreference.com/w/cpp/container/vector/resize) and [std::move_if_no_except](https://en.cppreference.com/w/cpp/utility/move_if_noexcept)